Automate p-tests
A Github CI workflow for a qemu ptest was added. The implementation is
simple, and is lacking in comprehensive exception handling. Rather than
commiting the log file to a repo, the results are being stored on the
build server (`/srv/test_results/ptest/<DATE>-ptest.log`). This was done
to (a) avoid adding extra steps... As the results will require
parsing/processing, this also avoids needing to clone the test logs
first.

Signed-off-by: Katrina Prosise <igorina@toganlabs.com>